Update patch set 1
Patch Set 1:

> Couldn't we just drop the module suffixes altogether, removing the
 > need for the symlinks?

How do you suggest we deal with the hardcoded object names, which are suffixed currently, but won't be in DDK 1.12 blobs?

 > About gralloc, yes, I think we should just use SGX core structure
 > for them as well.

We're completely sure that generated 4430/60 Gralloc blobs will always be the same, right? If the UM build system is the same as KM's, I guess no reason to expect otherwise.

 > I'd also consider moving the two different SGX core folders outside
 > proprietary/vendor/lib; what about something like,
 > vendor/ti/omap4/graphics/SGX.../?

Can do, once the Gralloc situation is clear.

Patch-set: 1