Update patch set 4
Patch Set 4:

> For instance, libIMGegl.so contains this:
 > PVRSRVLoadLibrary("/system/vendor/lib/egl/libGLESv1_CM_POWERVR_SGX540_120.so");
 
OK, that's definitely a good technical reason; it wasn't obvious when casually looking at the linked objects.

 > As you see, this isn't direct linking. We could just hex edit it as
 > well, but I'm more inclined towards using the naming scheme of the
 > newer blobs on jacinto6 (1.12+), or the older 1.8 blobs... but my
 > opinion can change if I get down-voted.

Yup, we definitely should avoid hex editing the newer libraries if we can help it. This was supposed to be a one time job for these ancient blobs, which are fixed now. I'm not sure why this inconsistency bothers me more than others; we could only rename them on device, but that would be confusing for purely cosmetic reasons.

It looks like we go with the current structure. Andi says he verified them too, but I didn't have a chance to test the 544 stuff.

Patch-set: 4